pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "toxic_comments"
version = "0.0.1"
description = "This project implements a transformer architecture with the purpose of classifying comments based on their toxicity."
authors = [
  { name = "Group 100", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc>=3.66.1",
    "dvc-gs>=3.0.2",
    "emoji>=0.6.0,<3.0.0",
    "fastapi==0.115.6",
    "hydra-core>=1.3.2",
    "lightning>=2.6.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pytorch-lightning>=2.6.0",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "streamlit>=1.53.0",
    "torch>=2.6.0",
    "transformers>=4.57.3",
    "typer==0.15.1",
    "uvicorn==0.40.0",
    "wandb>=0.23.1",
    "onnx>=1.20.1",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "scipy>=1.17.0",
    "google-cloud-storage>=3.8.0",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.14.11",
    "mkdocstrings-python==1.12.2",
    "snakeviz>=2.2.2",
    "pytest-mock>=3.10.0",
    "requests>=2.28.0",
    "httpx>=0.28.1",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
exclude = [
    "__init__.py",
]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["I", "E4", "E7", "E9", "F", "W", "D"]
ignore = ["D100", "D203", "D211", "D212", "D401"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.coverage.run]
omit = ["tests/*"]

[project.scripts]
train = "toxic_comments.train:main"
